Initial chunk reimplementation - tests don't pass yet

+++ core/src/world/mod.rs
@@ -104,11 +104,11 @@ impl ChunkMap {
     }
 }
 
-fn chunk_relative_pos(block_pos: BlockPosition) -> (u16, u16, u16) {
+fn chunk_relative_pos(block_pos: BlockPosition) -> (usize, usize, usize) {
     (
-        (block_pos.x % 16) as u16,
-        block_pos.y as u16,
-        (block_pos.z % 16) as u16,
+        (block_pos.x % 16) as usize,
+        block_pos.y as usize,
+        (block_pos.z % 16) as usize,
     )
 }
